move egress label setting of new nodes out of controller
